InitiateBuilderExit in gloas package
